RM: Make up a supervision tree for device deletion
Device deletion is asynchronous. As such, it may happen that the
RM service goes down during deletion, or that a temporary filure
brings down the deletion service. Therefore, a supervision tree
is put in place. The DeviceRemoverSupervisor supervises
Remover tasks, and the Reconciler handles the removal of
devices that may have been interrupted.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>