RM: Add device deletion logic
Device deletion begins when a device_id is put in the
`deletion_in_progress` table.
The actual deletion of data is started only when it has been
acknowledged by both DUP (two times) and the broker (once).
Deletion is started (or resumed, if needed) by the Scheduler
and it is carried out by a Task (see DeviceRemover).