DUP: handle device deletion
A DataUpdater process checks periodically if the related device
is being deleted. If that is the case, the process updates its
`deletion_in_progress` state subfield to `true` and writes
to db that deletion has started (`dup_start_ack` field).
From now on, all received messages will be acked but ignored,
in order to end consuming all inflight messages.
The DataUpdater also performs an RPC to the broker to
forcefully disconnect the device and prevent it from
reconnecting.
The last message to be processed by the DataUpdater
is by construction received on the internal `"/f"` path
(see https://github.com/astarte-platform/astarte_vmq_plugin/pull/75).
After that, the DUP deletion end ack is written to the database
and the DataUpdater process is terminated.

In order to allow deletion of offline (or old) devices,
the RemovalScheduler periodically checks the
`deletion_in_progress` table and makes the device process
start just to perform the deletion check.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>