Pairing: review consistency level of queries
The following assumption was made: at any given time, Astarte
should be up-to-date as to what regards basic realm entities
and configuration: therefore, device registration and update
queries (wwrites) have EACH_QUORUM consistency.
In case of a netsplit in the db, Astarte will thus be still
able to function, but no new entities will be added.
As a counterpart to this, read operations require LOCAL_QUORUM
consistency. This favours consistency over speed and availability,
but we err on the side of safety as Pairing is not a traffic
intensive component.
Finally, realm configuration (e.g. device registration limit)
is always agreed upon at any given time, so it is enough to read it
with ONE consistency.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>